+++ src/components/index.js
@@ -1,7 +1,7 @@
 export { default as Button } from './Button';
 export { default as Card } from './Card';
 export { default as ModalContainer } from './Modal';
-export { default as Navigation } from './Navigation';
+export { default as NavBar } from './NavBar';
 export { default as NotificationContainer } from './Notification';
 export { default as Content } from './Content';
 export { default as Counter } from './Counter';